+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Added new 'VerNumParser' class which provides utility functions for dealing with string-based version numbers.
+
 ## [v2.1.0] - 2018-01-22
 
 ### Added